src/tasks/task5 deployment/covid detection by Fahad Ahmed/requirements.txt
streamlit==1.2.0
numpy==1.21.4
tensorflow==2.7.0
Pillow==8.4.0